# Use faster copy when not overlapping