@klauspost 2019-09-01
This is measured on AMD64 but with `-tags=noasm`. I will add it to the description.